Add script nest level indicator
 #2

+++ install_scripts/curl.sh
@@ -54,7 +54,7 @@ else
     INSTALLED_VERSION=$(curl --version | head -n1 | cut -d' ' -f2 | sed 's/-DEV//')
   fi
 
-  if [ ! -z $REINSTALL ] || [ ! -f ${LOCAL_DIR}/bin/curl ]; then
+  if ([ $LEVEL = 0 ] && [ ! -z $REINSTALL ]) || [ ! -f ${LOCAL_DIR}/bin/curl ]; then
     iecho "$PKG_NAME $VER installation.. install location: $LOCAL_DIR"
 
     mkdir -p $TMP_DIR && cd $TMP_DIR
@@ -102,3 +102,4 @@ else
 
   cd $ROOT
 fi
+LEVEL=$(( ${LEVEL}-1 ))
